requirements.txt
requests>=2.31.0
websocket-client>=1.6.4
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2
schedule>=1.2.0
python-dateutil>=2.8.2
scipy>=1.11.4
